Moved to 'utils' folder. Applied new config
