+++ utils/httpClient.js
@@ -1,15 +1,15 @@
-const config = require("./config")();
+const config = require("./../config/server");
 
 async function GetOpensongRest(resource)
 {
-    let res = await fetch(`http://${config.openSongWsIp}:${config.openSongWsPort}/${resource}`);
+    let res = await fetch(`http://${config.opensongAddress}:${config.opensongPort}/${resource}`);
     return res.text()
 }
 
 async function PostOpensongRest(resource, payload = null)
 {
     let postOptions = {method: 'POST'};
-    let res = await fetch(`http://${config.openSongWsIp}:${config.openSongWsPort}/${resource}`, postOptions);
+    let res = await fetch(`http://${config.opensongAddress}:${config.opensongPort}/${resource}`, postOptions);
     return res.text()
 }
